requirements.txt
kafka-python
clickhouse-driver
Faker
psycopg2-binary
requests
cryptography
python-dotenv
fastapi
uvicorn
langchain
openai
httpx
langchain-community
aiofiles
aiokafka
asyncpg
# per OSRM-based routing nel producer
polyline

# per parsing GPX (se in futuro lo riutilizzi)
gpxpy

# per calcolo distanza haversine nel consumer
haversine

python-jose[cryptography]      # per JWT
passlib[bcrypt]               # per hashing password 
python-multipart

PyJWT
pydantic

redis==4.5.5    #sistema cache per llm


prometheus-client==0.17.1
prometheus-fastapi-instrumentator==6.1.0


faust-streaming==0.10.3
uvloop==0.18.0
structlog==23.1.0